requirements.txt
numpy
pandas
matplotlib
tensorflow
streamlit
sklearn
Pillow
seaborn
statsmodels
pandas-datareader
pmdarima
